feat: Created utility class for Time manipulation and formating
